requirements.txt
APScheduler==3.9.1
bcrypt==4.0.1
bidict==0.22.0
blinker==1.5
boto3==1.26.16
botocore==1.29.16
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
dominate==2.7.0
dropbox==11.35.0
email-validator==1.3.0
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
fonttools==4.37.4
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
MouseInfo==0.1.3
numpy==1.23.3
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221010
Pillow==9.2.0
ply==3.11
PyAstronomy==0.18.0
PyGetWindow==0.0.9
pymongo==4.3.2
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
quantities==0.13.0
requests==2.28.1
rubicon-objc==0.4.2
s3transfer==0.6.0
scipy==1.9.2
six==1.16.0
stone==3.3.1
tqdm==4.64.1
types-pytz==2022.4.0.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.1
